# swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Meal Compass API",
    "version": "2.0.0",
    "description": "REST API for planning weekly meals: users, recipes, meal plans and grocery lists with full CRUD.\n\n**Authentication:** log in by opening `/auth/github` in your browser (GitHub OAuth). The session cookie is then sent automatically, so \"Try it out\" works for protected routes. Without it, protected routes return 401. Users can only change their own data (403 otherwise); admins can change anything."
  },
  "servers": [
    {
      "url": "/",
      "description": "Current host (localhost or Render)"
    }
  ],
...
  "paths": {
...
    "/auth/me": {
      "get": {
        "tags": [
          "Auth"
        ],
        "summary": "Get the logged-in user",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Current user",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "401": {
            "description": "Not logged in",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/auth/logout": {
      "get": {
        "tags": [
          "Auth"
        ],
        "summary": "Log out (ends the session)",
        "responses": {
          "200": {
            "description": "Logged out",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "connect.sid",
        "description": "Session cookie set after GitHub login"
      }
    },
    "schemas": {
...
      "User": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string"
          },
          "githubId": {
            "type": "string",
            "example": "1234567"
          },
          "username": {
            "type": "string",
            "example": "andrea-g"
          },
          "displayName": {
            "type": "string",
            "example": "Andrea González"
          },
          "email": {
            "type": "string",
            "format": "email",
            "example": "andrea@example.com"
          },
          "avatarUrl": {
            "type": "string",
            "format": "uri",
            "example": "https://avatars.githubusercontent.com/u/1234567"
          },
          "role": {
            "type": "string",
            "enum": [
              "user",
              "admin"
            ],
            "default": "user",
            "description": "Only admins can change this."
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "lastLoginAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        }
      },
...
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "example": "Validation failed"
          },
          "details": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "field": {
                  "type": "string"
                },
                "message": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    }
  }
}